cloud_storage: move topic_manifest to a separate file
