+++ src/v/cloud_storage/CMakeLists.txt
@@ -4,6 +4,7 @@ v_cc_library(
   SRCS
     cache_service.cc
     manifest.cc
+    topic_manifest.cc
     recursive_directory_walker.cc
     remote.cc
     offset_translation_layer.cc
