+++ src/v/cloud_storage/partition_recovery_manager.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "cloud_storage/manifest.h"
 #include "cloud_storage/offset_translation_layer.h"
 #include "cloud_storage/remote.h"
+#include "cloud_storage/topic_manifest.h"
 #include "cloud_storage/types.h"
 #include "model/record.h"
 #include "s3/client.h"
